this is a shopping website for purchaging products
The name of this shopping website is
Green Horizzon which is an eco-friendly shopping website where you can buy products that are good for the environment. You can add items to your cart, buy them, or save them to a wishlist for later. There's also a login page where you can create an account to manage your orders and preferences. The website is designed using HTML and CSS for the front-end, making it easy to navigate, while MySQL is used for the back-end to store all the data securely.